requirements.txt
alembic==1.9.1
anyio==3.6.2
asgiref==3.6.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.88.0
googleapis-common-protos==1.56.2
grpcio==1.51.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-asgi==0.36b0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-instrumentation-sqlalchemy==0.36b0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
packaging==22.0
prometheus-client==0.15.0
prometheus-fastapi-instrumentator==5.9.1
protobuf==3.20.3
psycopg2==2.9.5
pydantic==1.10.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
thrift==0.16.0
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
